flask_backend/requirements.txt
Flask==2.1.1
Flask-Cors==3.1.1
Pillow==9.0.0
opencv-python==4.5.5.64
numpy==1.21.6
requests
ultralytics==8.0.39
